+++ DarkCodex/Components/AddFactOnlyParty.cs
@@ -1,8 +1,10 @@
 ﻿using Kingmaker;
 using Kingmaker.Blueprints;
 using Kingmaker.Blueprints.Classes.Selection;
+using Kingmaker.EntitySystem.Entities;
 using Kingmaker.PubSubSystem;
 using Kingmaker.UnitLogic;
+using Kingmaker.UnitLogic.Class.LevelUp;
 using Kingmaker.UnitLogic.Parts;
 using System;
 using System.Collections.Generic;
@@ -12,7 +14,7 @@ using System.Threading.Tasks;
 
 namespace DarkCodex.Components
 {
-    public class AddFactOnlyParty : UnitFactComponentDelegate
+    public class AddFactOnlyParty : UnitFactComponentDelegate, IUnitLevelUpHandler
     {
         public override void OnActivate()
         {
@@ -23,8 +25,17 @@ namespace DarkCodex.Components
             if (!isParty || this.Owner.HasFact(this.Feature))
                 return;
 
-            Helper.PrintDebug($"Applying {Feature.NameSafe()} for {this.Owner} is IsPet={Owner.IsPet} IsMainCharacter={Owner.IsMainCharacter} IsCustomCompanion={Owner.IsCustomCompanion()} IsStoryCompanion={Owner.IsStoryCompanion()}");
+            Helper.Print($"Applying {Feature.NameSafe()} for {this.Owner} is IsPet={Owner.IsPet} IsMainCharacter={Owner.IsMainCharacter} IsCustomCompanion={Owner.IsCustomCompanion()} IsStoryCompanion={Owner.IsStoryCompanion()}");
             this.Owner.AddFact(this.Feature, null, this.Parameter);
+        }
+
+        public void HandleUnitBeforeLevelUp(UnitEntityData unit)
+        {
+            OnActivate();
+        }
+
+        public void HandleUnitAfterLevelUp(UnitEntityData unit, LevelUpController controller)
+        {
         }
 
         public BlueprintUnitFactReference Feature;